pushbullet-notifications/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pushbullet.py==0.9.1
python-magic==0.4.15
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
